getContact implemented.
getContact has been implemented for both the box and circle collider using a contact struct. The contact methods determin whether the two objects overlap and create a normal vector which makes it so the two objects no longer overlap. It will then return the contact object so that another method can use it to actually move the two objects apart.